Issue #3804
more review changes
